data-engineering/requirements.txt
paho-mqtt>=2.0
python-dotenv>=1.0
requests>=2.31
taospy>=2.7
pymysql>=1.1
